include: make ParallelConfig::restore_pc_from_str safe
ParallelConfig was probably a POD struct at some point, but when a string
was added, it became undefined/unsafe to memset.
Also added assertions to check that assigned counter fields are in bounds
for max array length.

It should also be noted that the string field is not thread-safe in
ParallelConfig::get_pc_str.